requirements.txt
yfinance
python-telegram-bot==13.15
matplotlib
